+++ src/utils/vec_dot.c
@@ -6,17 +6,17 @@
 /*   By: jeepark <jeepark@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/16 16:45:58 by jeepark           #+#    #+#             */
-/*   Updated: 2022/09/21 15:00:53 by jeepark          ###   ########.fr       */
+/*   Updated: 2022/09/25 16:40:03 by jeepark          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "mini_rt.h"
 
-float   vec_dot(t_vec3 v1, t_vec3 v2)
+double   vec_dot(t_vec3 v1, t_vec3 v2)
 {
-    float   x;
-    float   y;
-    float   z;
+    double   x;
+    double   y;
+    double   z;
 
     x = v1.x * v2.x;
     y = v1.y * v2.y;
